+++ CHANGELOG.md
@@ -1,6 +1,12 @@
-# lita-nexus change log
+# Change Log
+
+## 0.1.2
+* Allow to change/show current working repository
+* Allow to delete a repository
+* Allow to delete an artifact with coordinate
 
 ## 0.1.1
 Fix help message.
+
 ## 0.1.0
 Initial release.